requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alpha_vantage==3.0.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cftime==1.6.4.post1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cpp==1.0.0.6
cycler==0.12.1
distro==1.9.0
filelock==3.20.3
Flask==3.1.0
fonttools==4.56.0
frozenlist==1.6.0
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
lightgbm==4.6.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
multidict==6.4.3
narwhals==1.37.1
netCDF4==1.7.2
networkx==3.6.1
nltk==3.9.2
numpy==2.2.3
openai==2.3.0
opencv-python==4.13.0.92
packaging==24.2
pandas==2.2.3
pillow==11.1.0
propcache==0.3.1
protobuf==6.30.2
py4j==0.10.9.7
pyarrow==20.0.0
pydantic==2.12.0
pydantic_core==2.41.1
pydeck==0.9.1
pyparsing==3.2.1
pyspark==3.5.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.1
referencing==0.36.2
regex==2025.9.18
requests==2.32.3
rpds-py==0.24.0
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
setuptools==82.0.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.50.0
sympy==1.14.0
tenacity==9.1.2
textblob==0.19.0
threadpoolctl==3.5.0
toml==0.10.2
torch==2.10.0
torchvision==0.25.0
tornado==6.4.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.1
urllib3==2.4.0
watchdog==6.0.0
Werkzeug==3.1.3
xarray==2025.1.2
yarl==1.20.0